Move notification package to lib/ and add open package
