+++ lib/open/doc.go
@@ -1,0 +1,4 @@
+// Package open contains a simple cross-platform way to open files in the program the OS wants to use.
+//
+// Based on https://github.com/skratchdot/open-golang
+package open
